Ensure zellij does not run inside vscode integrated term

+++ dotfiles/fish/.config/fish/config.fish
@@ -18,39 +18,41 @@ if status is-interactive
     end
 
     # enable zellij integration
-    set -x ZELLIJ_AUTO_ATTACH true
-    set -x ZELLIJ_AUTO_EXIT true
+    if test "$TERM_PROGRAM" != vscode
+        set -x ZELLIJ_AUTO_ATTACH false
+        set -x ZELLIJ_AUTO_EXIT false
 
-    if not set -q ZELLIJ
-        if test "$ZELLIJ_AUTO_ATTACH" = true
-            zellij attach -c
-        else
-            zellij
-        end
+        if not set -q ZELLIJ
+            if test "$ZELLIJ_AUTO_ATTACH" = true
+                zellij attach -c
+            else
+                zellij
+            end
 
-        if test "$ZELLIJ_AUTO_EXIT" = true
-            kill $fish_pid
-        end
-    end
-
-    if type -q zellij
-        # Update the zellij tab name with the current process name or pwd.
-        function zellij_tab_name_update_pre --on-event fish_preexec
-            if set -q ZELLIJ
-                set -l cmd_line (string split " " -- $argv)
-                set -l process_name $cmd_line[1]
-                if test -n "$process_name" -a "$process_name" != z
-                    command nohup zellij action rename-tab $process_name >/dev/null 2>&1
-                end
+            if test "$ZELLIJ_AUTO_EXIT" = true
+                kill $fish_pid
             end
         end
 
-        function zellij_tab_name_update_post --on-event fish_postexec
-            if set -q ZELLIJ
-                set -l cmd_line (string split " " -- $argv)
-                set -l process_name $cmd_line[1]
-                if test "$process_name" = z
-                    command nohup zellij action rename-tab (prompt_pwd) >/dev/null 2>&1
+        if type -q zellij
+            # Update the zellij tab name with the current process name or pwd.
+            function zellij_tab_name_update_pre --on-event fish_preexec
+                if set -q ZELLIJ
+                    set -l cmd_line (string split " " -- $argv)
+                    set -l process_name $cmd_line[1]
+                    if test -n "$process_name" -a "$process_name" != z
+                        command nohup zellij action rename-tab $process_name >/dev/null 2>&1
+                    end
+                end
+            end
+
+            function zellij_tab_name_update_post --on-event fish_postexec
+                if set -q ZELLIJ
+                    set -l cmd_line (string split " " -- $argv)
+                    set -l process_name $cmd_line[1]
+                    if test "$process_name" = z
+                        command nohup zellij action rename-tab (prompt_pwd) >/dev/null 2>&1
+                    end
                 end
             end
         end
